aom_fw-222 Need to store application data in database

+++ Cache/System.php
@@ -92,15 +92,27 @@ namespace Aomebo\Cache
                 \Aomebo\Configuration::getSetting('cache,expiration time')
              ) {
 
-                 if (\Aomebo\Database\Adapter::query(
-                     'DELETE FROM `' . self::getTable() . '` '
-                     . 'WHERE `cache_added` < NOW() - INTERVAL {time} SECOND',
-                     array(
-                         'time' => $cacheExpiration,
-                     ))
-                 ) {
-                     return true;
-                 }
+                $lastCheck =
+                    \Aomebo\Application::getApplicationData('last_cache_garbage_collect');
+
+                if (!isset($lastCheck)
+                    || $lastCheck < time() - 86400
+                ) {
+
+                    \Aomebo\Database\Adapter::query(
+                         'DELETE FROM `' . self::getTable() . '` '
+                         . 'WHERE `cache_added` < NOW() - INTERVAL {time} SECOND',
+                         array(
+                             'time' => $cacheExpiration,
+                         )
+                    );
+
+                    \Aomebo\Application::setApplicationData(
+                        'last_cache_garbage_collect',
+                        time()
+                    );
+
+                }
 
              }
